More Norminette - Everything is norme proof (apart ft_bfs and the files taht will be deleted) - The only pending is to delete all comments from everywhere - Also changed some filenames
