+++ lemin_files/ft_hashing.c
@@ -6,7 +6,7 @@
 /*   By: dsaripap <dsaripap@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/06/19 10:38:04 by dsaripap      #+#    #+#                 */
-/*   Updated: 2020/07/29 14:46:30 by dsaripap      ########   odam.nl         */
+/*   Updated: 2020/07/31 05:26:22 by dsaripap      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -92,6 +92,32 @@ unsigned int		ft_hash_function(char *name, int size)
 }
 
 /*
+** Function that is executed when a colision happens in the Hash Table.
+** We handle colisions with the the separate chaining method
+** so in the same key of the hash table we can have more than
+** one room stored.
+*/
+
+void				ft_hashing_colision(t_ant_farm *ant_farm, \
+									t_room *temp, unsigned int key)
+{
+	t_hash_item		*ht_item;
+	t_hash_item		*temp_item;
+
+	// ft_printf(" Key '%d' exists\n", key);
+	temp_item = ant_farm->hash_table->array[key];
+	// while (temp_item->colision_next != NULL)
+	// {
+	// 	ft_printf(" After item %s of the collision list\n", temp_item->room_name);
+	// 	temp_item = temp_item->colision_next;
+	// }
+	// ft_printf("Colision : Saving room '%s' in the linked list of key '%d'\n", temp->name, key);
+	ht_item = ft_hashitem_newnode(temp, key);
+	// ht_item->hashed_key = key;
+	ft_hashitem_addend(&(temp_item->colision_next), ht_item);
+}
+
+/*
 ** Function that saves a room in the Hash Table
 ** For every room we malloc a hash_item (which is an item in the hash_table)
 ** And then to save the room_name in the hash_item->room_name
@@ -105,7 +131,7 @@ int					ft_hashing_process(t_ant_farm *ant_farm)
 {
 	unsigned int	key;
 	t_hash_item		*ht_item;
-	t_hash_item		*temp_item;
+	// t_hash_item		*temp_item;
 	t_room			*temp;
 
 	temp = ant_farm->rooms_lst;
@@ -126,19 +152,20 @@ int					ft_hashing_process(t_ant_farm *ant_farm)
 			ft_hashitem_addend(&(ant_farm->hash_table->array[key]), ht_item);
 		}
 		else
-		{
-			// ft_printf(" Key '%d' exists\n", key);
-			temp_item = ant_farm->hash_table->array[key];
-			// while (temp_item->colision_next != NULL)
-			// {
-			// 	ft_printf(" After item %s of the collision list\n", temp_item->room_name);
-			// 	temp_item = temp_item->colision_next;
-			// }
-			// ft_printf("Colision : Saving room '%s' in the linked list of key '%d'\n", temp->name, key);
-			ht_item = ft_hashitem_newnode(temp, key);
-			// ht_item->hashed_key = key;
-			ft_hashitem_addend(&(temp_item->colision_next), ht_item);
-		}
+			ft_hashing_colision(ant_farm, temp, key);
+		// {
+		// 	// ft_printf(" Key '%d' exists\n", key);
+		// 	temp_item = ant_farm->hash_table->array[key];
+		// 	// while (temp_item->colision_next != NULL)
+		// 	// {
+		// 	// 	ft_printf(" After item %s of the collision list\n", temp_item->room_name);
+		// 	// 	temp_item = temp_item->colision_next;
+		// 	// }
+		// 	// ft_printf("Colision : Saving room '%s' in the linked list of key '%d'\n", temp->name, key);
+		// 	ht_item = ft_hashitem_newnode(temp, key);
+		// 	// ht_item->hashed_key = key;
+		// 	ft_hashitem_addend(&(temp_item->colision_next), ht_item);
+		// }
 		temp = temp->next;
 		// print_hash_table(ant_farm->hash_table);
 	}
