+++ lemin_files/ft_sort_neighbors.c
@@ -6,7 +6,7 @@
 /*   By: dsaripap <dsaripap@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/06/19 10:39:48 by dsaripap      #+#    #+#                 */
-/*   Updated: 2020/07/21 12:02:15 by dsaripap      ########   odam.nl         */
+/*   Updated: 2020/07/30 11:07:43 by dsaripap      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,46 +63,64 @@ void				ft_splitlist(t_neighbor *completelist, \
 	}
 }
 
+// static void			ft_set_result(t_neighbor **result_list, \
+// 								t_neighbor **a, t_neighbor **b, \
+// 								size_t flag)
+// {
+// 	(*result_list) = *a;
+// 	result_list->next = ft_mergeback(&((*a)->next), b, flag);
+// 	result_list->next->prev = *a;
+// }
+
 /*
 ** The merging part of merge sort
 */
 
-t_neighbor			*ft_mergeback(t_neighbor **ahalf, t_neighbor **bhalf, size_t flag)
+t_neighbor			*ft_mergeback(t_neighbor **ahalf, t_neighbor **bhalf, \
+									size_t flag)
 {
 	t_neighbor		*result_list;
 
 	result_list = NULL;
+	// (*ahalf != NULL) ? : return (*bhalf);
+	// (*bhalf != NULL) ? : return (*ahalf);
 	if (*ahalf == NULL)
 		return (*bhalf);
 	else if (*bhalf == NULL)
 		return (*ahalf);
+	// else
+	// {
+	// if (((*ahalf)->hash_item->room != NULL) && ((*bhalf)->hash_item->room != NULL))
+	// 	ft_printf("ahalf %s = %d, bhalf %s = %d \n", (*ahalf)->hash_item->room->name, (*ahalf)->hash_item->room->score, (*bhalf)->hash_item->room->name, (*bhalf)->hash_item->room->score);
+	// if ((*ahalf)->hash_item->room->score < (*bhalf)->hash_item->room->score)
+	if ((flag == 0 && (*ahalf)->hash_item->room->score > \
+	(*bhalf)->hash_item->room->score) || \
+	(flag == 1 && (*ahalf)->hash_item->room->level_sink > \
+	(*bhalf)->hash_item->room->level_sink))
+	{
+		// ft_set_result(&result_list, bhalf, ahalf, flag);
+		// ft_printf("bhalf first \n");
+		result_list = *bhalf;
+		result_list->next = ft_mergeback(&((*bhalf)->next), ahalf, flag);
+		result_list->next->prev = *bhalf;
+	}
+	// else if (flag == 1 && (*ahalf)->hash_item->room->level_sink > \
+	// (*bhalf)->hash_item->room->level_sink)
+	// {
+		// ft_printf("bhalf first \n");
+		// ft_set_result(&result_list, bhalf, ahalf, flag);
+		// result_list = *bhalf;
+		// result_list->next = ft_mergeback(&((*bhalf)->next), ahalf, flag);
+		// result_list->next->prev = *bhalf;
+	// }
 	else
 	{
-		// if (((*ahalf)->hash_item->room != NULL) && ((*bhalf)->hash_item->room != NULL))
-		// 	ft_printf("ahalf %s = %d, bhalf %s = %d \n", (*ahalf)->hash_item->room->name, (*ahalf)->hash_item->room->score, (*bhalf)->hash_item->room->name, (*bhalf)->hash_item->room->score);
-		// if ((*ahalf)->hash_item->room->score < (*bhalf)->hash_item->room->score)
-		if (flag == 0 && (*ahalf)->hash_item->room->score > (*bhalf)->hash_item->room->score)
-		{
-			// ft_printf("bhalf first \n");
-			result_list = *bhalf;
-			result_list->next = ft_mergeback(&((*bhalf)->next), ahalf, flag);
-			result_list->next->prev = *bhalf;
-		}
-		else if (flag == 1 && (*ahalf)->hash_item->room->level_sink > (*bhalf)->hash_item->room->level_sink)
-		{
-			// ft_printf("bhalf first \n");
-			result_list = *bhalf;
-			result_list->next = ft_mergeback(&((*bhalf)->next), ahalf, flag);
-			result_list->next->prev = *bhalf;
-		}
-		else
-		{
-			// ft_printf("ahalf first \n");
-			result_list = *ahalf;
-			result_list->next = ft_mergeback(&((*ahalf)->next), bhalf, flag);
-			result_list->next->prev = *ahalf;
-		}
+		// ft_printf("ahalf first \n");
+		result_list = *ahalf;
+		result_list->next = ft_mergeback(&((*ahalf)->next), bhalf, flag);
+		result_list->next->prev = *ahalf;
 	}
+	// }
 	return (result_list);
 }
 
@@ -136,8 +154,8 @@ t_neighbor			*ft_merge_sort(t_neighbor **completelist, size_t flag)
 ** Checks which room has the smallest score and puts that first
 */
 
-int					ft_sort_neighbors(t_neighbor **completelist, t_room **room, \
-									size_t flag)
+int					ft_sort_neighbors(t_neighbor **completelist, \
+									t_room **room, size_t flag)
 {
 	(*room)->neighbors = ft_merge_sort(completelist, flag);
 	// ft_printf(" after merge %s", (*room)->neighbors->hash_item->room->name);
