+++ lemin_files/lemin.c
@@ -6,7 +6,7 @@
 /*   By: dsaripap <dsaripap@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/06/19 10:39:57 by dsaripap      #+#    #+#                 */
-/*   Updated: 2020/07/29 19:27:43 by dsaripap      ########   odam.nl         */
+/*   Updated: 2020/07/31 05:55:48 by dsaripap      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,17 +59,17 @@ int					main(int argc, char **argv)
 	if (ft_save_neighbors(ant_farm) != SUCCESS)
 		return (ft_exitprogram(ant_farm));
 	// ft_print_neighbors_list(ant_farm->hash_table);
-	// ft_print_rooms_list(ant_farm->rooms_lst);
 	// delete_dead_ends(ant_farm->hash_table);
 	// ft_printf("Amount of rooms : %d\n", ant_farm->rooms);
 	ft_bfs_level_sink(ant_farm);
 	ft_bfs_level_source(ant_farm);
+	ft_print_rooms_list(ant_farm->rooms_lst);
 	ft_bfs_runs(ant_farm);
 	ft_free_paths_ants_lst(ant_farm);
 	ft_ants_to_paths(ant_farm);
-	ft_printf(ANSI_COLOR_BLUE" ============================================= \n");
-	ft_printf(" ****** BEST RUN %d - Lines needed %d ********** \n", ant_farm->best_run, ant_farm->lines);
-	ft_printf(" ============================================= \n"ANSI_COLOR_RESET);
+	// ft_printf(ANSI_COLOR_BLUE" ============================================= \n");
+	// ft_printf(" ****** BEST RUN %d - Lines needed %d ********** \n", ant_farm->best_run, ant_farm->lines);
+	// ft_printf(" ============================================= \n"ANSI_COLOR_RESET);
 	// ft_print_paths(ant_farm);
 	// ft_print_paths_sizes(ant_farm);
 	// ft_bfs_again(ant_farm, 2);
@@ -90,7 +90,7 @@ int					main(int argc, char **argv)
 	// else
 	// 	ft_printf("its set %d\n", ant_farm->options & OPTION_M);
 	ft_move_ants(ant_farm);
-	ft_print_paths_list(ant_farm);
+	// ft_print_paths_list(ant_farm);
 	// ft_print_paths_list_detail(ant_farm);
 	ft_exitprogram(ant_farm);
 	// while (1)
